Format, capsule and WIP

+++ backend/src/TippingContract.sol
@@ -1,35 +1,31 @@
 pragma solidity ^0.8.18;
 
-import "openzeppelin-solidity/contracts/token/ERC721/IERC721.sol";
-import "openzeppelin-solidity/contracts/token/ERC20/IERC20.sol";
+import "openzeppelin-contracts/contracts/token/ERC721/IERC721.sol";
+import "openzeppelin-contracts/contracts/token/ERC20/IERC20.sol";
 
 contract TippingContract {
-
-struct token {
+    struct token {
         uint40 timestamp;
         address owner;
-}
+    }
 
     IERC721 public momeNFT;
     IERC20 public apeCoin;
     mapping(uint256 => uint256) public tipAmounts; // tokenId => amount
     mapping(uint256 => token) public tokens; // tokenId => token
-    
 
     constructor(address _momeNFT, address _apeCoin) {
         momeNFT = IERC721(_momeNFT);
         apeCoin = IERC20(_apeCoin);
     }
 
-    function tip(uint256 tokenId, uint amount) public {
+    function tip(uint256 tokenId, uint256 amount) public {
         tipAmounts[tokenId] += amount;
         apeCoin.transferFrom(msg.sender, address(this), amount);
         momeNFT.transferFrom(address(this), msg.sender, tokenId);
     }
 
-    function registerToken(uint256 tokenId) public {
-        
-    }
+    function registerToken(uint256 tokenId) public {}
 
     function withdraw(uint256 tokenId) public {
         uint256 amount = tipAmounts[tokenId];
@@ -38,4 +34,3 @@ struct token {
         momeNFT.transferFrom(msg.sender, address(this), tokenId);
     }
 }
-
